Updating location primitive

+++ ClientSDKs/BeerDrinkin.DataObjects/CheckIn.cs
@@ -5,8 +5,8 @@
         public string BeerId { get; set; }
         public string UserId { get; set; }
         public string[] Images { get; set; }
-        public float Longitude { get; set; }
-        public float Latitude { get; set; }
+        public double Longitude { get; set; }
+        public double Latitude { get; set; }
         public bool IsBottled { get; set; }
         public string RatingId { get; set; }
